services/task-service/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
